+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Changelog
 Development
 -----------
 
-* Placeholder
+* Add `acc_electricity_utilities_climate_zones` table for easy access.
+* Rename `deer_ls` to `deer_load_shapes`.
 
 0.0.1
 -----